Fixed a crash when trying to switch servers
Ownership of the charInfo global variable wasn't well defined. It was
being locked, unlocked and generally modified from a lot of places, and
somewhere in this mess it ended up crashing when switching servers.

Now the CharHandler instances, for eAthena and manaserv respectively,
own this list of characters. A new class, Net::Character wraps up the
slot index in combination with the player dummy. The list is passed on
to the CharSelectDialog each time it changes.

Both related and unrelated cleanups were made as well.

+++ src/net/ea/playerhandler.cpp
@@ -556,7 +556,7 @@ void PlayerHandler::emote(int emoteId)
     outMsg.writeInt8(emoteId);
 }
 
-void PlayerHandler::increaseAttribute(size_t attr)
+void PlayerHandler::increaseAttribute(int attr)
 {
     if (attr >= STR && attr <= LUK)
     {
@@ -566,7 +566,7 @@ void PlayerHandler::increaseAttribute(size_t attr)
     }
 }
 
-void PlayerHandler::decreaseAttribute(size_t attr)
+void PlayerHandler::decreaseAttribute(int attr)
 {
     // Supported by eA?
 }
