+++ src/net/playerhandler.h
@@ -35,9 +35,9 @@ class PlayerHandler
 
         virtual void emote(int emoteId) = 0;
 
-        virtual void increaseAttribute(size_t attr) = 0;
+        virtual void increaseAttribute(int attr) = 0;
 
-        virtual void decreaseAttribute(size_t attr) = 0;
+        virtual void decreaseAttribute(int attr) = 0;
 
         virtual void increaseSkill(int skillId) = 0;
